requirements.txt
openai
tiktoken
ragas
python-dotenv
huggingface_hub
qdrant_client
fastembed
pytest

langchain
langchain_experimental
langchain-anthropic
langchain_huggingface
langchain-openai

llama_index==0.10.36
llama-index-vector-stores-qdrant==0.2.10
llama-index-embeddings-fastembed==0.1.4
llama-index-multi-modal-llms-openai
llama-index-llms-anthropic
llama-index-llms-openai

pytest